Refactored UART driver into low-level and high-level parts
low-level driver will be shared by main UART ports (A-C) and the
keyboard/terminal driver.

Removed DEVICE_DRIVER pointer as it was a pest to setup for the uart driver
and no real use anyway. Tided up some labels in various places to make
them local and not global (more to do there)