Two big changes: Console driver, and the image is now linked
The console driver uses the uartlowlevel "library". Includes scancode
to ASCII translation. Test task1 now uses the keyboard.

Also, instead of using horrible .includes to tie the thing together,
the sources are now assembled and link as two steps, using aslink as it
was meant to be used. A list of global symbols, the OS's "API", is
generated with a special Makefile target.

Three .areas are used: ROM, RAM and VECTORS. The base address for these
are set on the aslink commandline. main is forced to the front so that
it .orgs to 0xc000.

TODO: more tidying! Subdirectories for drivers and includes would be
nice.