Improvements to task display in monitor
Shows most fields from the task struct, including registers last stacked.

While testing this I realised that TASK_SIAGALLOC was not being updated
when the default IO device was moved to the created child. Fix is a
bit of a bodge: free the old signal bit and move it to 0x01 in the new
task.