Made monitor commands words instead of letters
parambuffer has been extended to contain the command, followed by a null.
This is then compared against each string in the command table using
strcmp.

Also added some more comments to the monitor code, though it needs still
more.

I might make the commandarray mechanism into an API which would be called
with the command array and the entered command, and return a matching
function pointer. Not sure yet if this would be useful from a
CLI/Shell though, since the parambuffer format doesn't really work
for a shell used for running high level commands...