Switch enable to permit is most places
During some criticial sections, eg when allocating memory, it is only
necesary to stop task switching, not all interrupt handlers. This is
because an ISR is not allowed to mess with the memory list so interrupts
can be dealt with even when user code is playing with the memory list.

This slightly reduces the chance of data loss (eg. on UARTs) while
task switching is disabled.