Shell improvements
Added a new option type to the parser, type=4. The byte following is the
"option" character, which is parsed from strings starting -. eg:

list -l

Which parses out to 0x04 followed by whatever ASCII l is.

This is used to "list long" the current directory, which shows info from
the inode, allbeit in hex.

Fixed a minor problem with the parer which prevented a null being inserted
if the command did not have any options, which caused repeated calls to the
parser with different things following the command to go a bit goofy.