Added an initial joystick driver
Very simple: unit 0 for port 0 and unit 1 for port 1.

Eventually I'll extend the joystick driver to have an event mode, which will
use a blocking read until a change is detected by a timer interrupt handler.
To do this I need to find the best way to represent a change on one of the
five inputs in the read-in byte...

Fixed a "success/fail" error in the LED driver.