Added xrun command for xmodem-ing a subroutine file
This involved fixing copious little bugs, including subtle rola vs lsla
problems in the uart driver. xrun takes a memory size, allocates that
many bytes, opens uart port b at 19200 and rx the file using xmodem.
Finally it closes the port then runs the block as a subroutine, setting
up x as the shell's io channel. On rts it prints the a reg so the sub
can return a value.

Most of the issues were around the uart open/close/open/close which had
not been done in any previous code.

Tidied up the uart isr by making it check that the port is actually open
before it tries to read the uart registers. Added a few more debugs.

Also in lib/io.asm I added getchardefio and putchardefio for getting a
byte on the defaultio. Also in here I renamed "bytes" to "chars" when the
routine was not dealing with hex but actual raw values.

The "API" to user code, as used by xrun is a inc file autogenerated
from the map. Currently it includes all globals. I'm unsure what this
will eventually look like, but right now its a PITA having to rebuild
external code everytime I change the ROM.

Added my as6809 wrapper into tools. It deletes the .rel on a build failure
to stop make getting confused.

Included a sample external routine which does stuff with the IO channel.