Make loader honour the reset vector at 0xfffe
Instead of booting into a hardcoded 0xc000, boot into the reset vector
held in the EEPROM. This entails always copying the loader into RAM and
then jumping through the copy, which allows us to expose the top page
of EEPROM. Previously we only copied the loader into RAM if we were flashing
but now, because we need to read the EEPROM vector inside the loader, we
have to always do the copy operation.

This was tesed by inserting padding into the first page of the EEPROM,
to move the reset vector along abit from 0xc000. Soon (maybe the next
commit) an external global vector table will go in the bottom of the
EEPROM.

This also has the benifit of allowing booting into any EEPROM with a
properly configured reset vector, instead of only starting from 0xc000.