Added automatic creation of the subtable
The subtable (subroutine table) sits at the bottom of ROM, from 0xc000. The
table is an array of .words for all the in ROM non underscore-starting
globals.  To update the table, run "make subtable" which will add new
entries (as found in the map) to the bottom of main/subtable.asm.  This .rel
from this .asm is always at the bottom of ROM, and main/main.rel can go now
go anywhere, thanks to the change to the loader to get it read the reset
vector.

The subtable is not automatically created, since it needs the map to be
present. Thus to do a clean build from scratch you need to do one build to
get the list of vectors, and a further build once the table has been made
to include that table in the output. Note that the table is a table of
symbol references, not the actual routine locations, so the table is
actually filled out at link time. While the table could have been worked
out by processing the .asm files, this would be more haphazard (probs)
then processing the .map.

There are thus two "user includes":

systemvars.inc: made always at build and contains equates for RAM globals
(all of which are in main/systemvars.asm). I may trim this list at some
point, since it currently includes internal stuff like isr addresses.

subtable.inc: made on demand with "make subtable". This contains the system
routine vectors.

The test sub .asm has been updated to call the routines using indirect
addressing. This code should no longer need rebuilding with ROM layout
changes, so long as the subtable stays at the bottom of ROM and does not
change its ordering.