Added an SPI driver, and tidy ups
Debug builds were broken since fiddling with hidden globals, so fixed them.
Also found out what the spacez gobal was used for (debug). Made that a
private global.

SPI is bit-banged through DISCo, but one improvement over the previous
monitor-OS: sysopen "ide" UNIT, where UNIT gets translated to different
slave select vales by logic in DISCo.  0=Joy0, 1=Joy1, 2=DS1305, 3=CAT25256.

Driver is simple syswrite/sysread affair with syscontrol for making a new
transaction. Tested with the monitor and it seems to work well.

gettime and settime are two user programs which get and set the time. These
are ports of the old code, with some minor improvmenets. Added string to
byte routine to lib/strings as an external.

Tidy up of xrun command to move the actual sub call into its own sub, so
that u can be stacked.  It holds the base address and it cannot be modified
by the user subroutine so it is explicitly stacked.