Moved the console video RAM usage under 0x8000
This will leave the high 32KB free for non text stuff. An F-key, probably
F10, will be used to switch to this "alternative" mode.  From there it will
be possible to switch back from inside the game (etc).  I have not yet
figured out how keyboard access will work inside this "alternative" useage
of the VDC, but the rough idea is it should be possible to make simple games
like my Snake game which run as tasks without doing away with the task
switching and driver model of the OS.