Added IO_ERR_EOF, for when sysread has reached the end of the file
Checking EOF now happens at the top of miniread, so the last char in the
file will always be returned. The *next* char after the last will trigger
IO_ERR_EOF.

Reworked lib/io to use getchar and putchar everywhere, which calls
handleioerror when sysread/syswrite return notzero. a will have the
error, which is one of wait, break or eof. Wait and break are "non fatal"
errors, so the caller (say gethcar) will retry. This means the wait logic
is all in handleioerror, making getstr simpler.

getnextdirent also follows this logic; errors are dealt with by returning
with notzero instead of the caller looking for an empty dirent.