Added a bodge so user .bins contain bss data
Previously the .bin files did not hold the .rmb bytes for input buffers
etc. This presents a problem with the .bins are read in to freshly allocated
memory and run: the .rmb buffers are not inside the allocated memory, and
thus corruption occurs on following allocated memory blocks (probably the
big free block at the end of the list)

Work around this by forcing a single byte onto the end of the object code.
This will cause the .bin file to always be big enough so that the .rmb
data is in allocated space.

The proper fix for this to add a header to the external executables, which
includes things like the size of memory to allocate, in addition to the
actual code.

(Also switch from using hex2bin to standard objcopy, which I think is in
binutils. This was also done to the main MAXI09OS EEPROM)