Added the Snake game, a "port" to MAXI09OS
This is hacky, but it works!

TODO: Currently it busyloops, and does not use any OS drivers.

Uses the new setgraphicsub call in the console driver to set a callback for
the F10 graphial screen.

You can happily switch terminals or use the serial port while the game
runs, and it even recovers the screenmode when the game exits.