Added fileserver tool and refactored serial coms into standalone obj
Fileserver is used by the getrun shell call, which obtains the content
of the given file over UART port 1 and runs it as a sub. This allows
external programs to be run without copying them to the Compact Flash.

The protocol is, naturally trivial:

Linux sends: <command byte><filename>\0
6809 sends (for get command): <response byte><length word><file data>

Respones are sent for things like file not found, but these are not
currently displayed to the user.

For performance reasons, multitasking is disabled while the transfer
happens. I did play (again) with getting hardware handshaking working,
but failed, probably because of issues on the Linux end.