Added getrun shell command, and associated stuff
Change overflow detection in UART driver to use a "fill level" for the
circ buffer, instead of trying to monkey about with the offsets. Debug-
log when an overflow occures between sysread and the ISR, but do nothing
about it.

See the commit in tools/ made about the same time for info on getrun.