Added initial work for writing user programs in C(!)
Usses the 6809 gcc port currently, though porting to cmoc would not be
hard.

Binaries (ubuntu but they work in debian) for tools is here:
https://launchpad.net/~tormodvolden/+archive/ubuntu/m6809

The "C library" is shockingly minimal currently. printf comes from the
mini-printf project on github: https://github.com/mludvig/mini-printf -
obviously this is a static linkage, since MAXI09OS sdoes not support shared
modules (yet?)

It also includes a altermative (vs the one in gcc-6809) integer division
implementation, which I found here:
https://storage.googleapis.com/google-code-archive-downloads/v2/code.google.com/gcc6809/div32.s

This was because the one in libgcc seems to cause crashes.