Add support for building with CMOC via a variable in Makefile.inc
CMOC is another C compiler, written specifically for the 6809:
https://perso.b2b2c.ca/~sarrazip/dev/cmoc.html

It has a few missing C features, but it doesn't suffer from at least one
gcc6809 problem: with -fPIC enabled, access to pointers which are
global/static is broken: no account for the "position independant" nature of
the content behind the pointer. PC relative addressing is used to get the
poiner, but not the thing being pointed too.

(Actually with the lwtools gcc6809 version, even the access to the base of
the array is missing the ,pcr addressing mode.  I had to do a build from
github to produce the ,pcr mode for the base of the array, but it still
leaves the problem of the pointers themselves.  Modern CPUs/compilers solve
this with a Glboal Offset Table, which is implemented with a register which
holds an offset for each global)

So I've been looking at CMOC. Other then missing language features (eg. no
"const") the biggest problem is performance: the code it makes is about half
the speed of gcc code. On the oher hand there is no bodgering about with
linker scripts or custom integer divison routines. It is possible to either
use lwlink directly or use cmoc to do the linking. cmoc seems to be preffered
because it will strip unneeded code, although lwlink is more "native"

Also, I've implemented more functions in libmaxi09os. To test them out I've
written gettime in C. It's nicer to read, but bigger and slower then the
assembly version!