Switch the SPI driver to doing shifting using DISCo
Shifting has been implemented (but probably needs a bit of a cleanup) in
DISCo. syswrite will also return the last read in byte, due to how
some devices, like the ADC in MAXI09, are bit orientated and do not
split data on bytes.

nops are used to pause while the byte is shifted, instead of using
a status bit. This will be implemented some time in the future

This hardware approach to SPI should be much faster then the old
code where the shifting is done in the MPU..