Made the hl isr for the uart port a property of the lll port
The handler is now saved in the low level code, so different high level uart
users (the normal uart driver and the console driver atm) set the handler
pointer on the low level open.  Eventually this will be used for the SLIP(!)
driver which will have a packet based task signal mechanism instead of a
byte-orientated one.

Also added a control call to the console driver, which is used to send
a command byte to the keyboard MCU so the RGB LED can be fiddled with, and
the key repeat speeds changed.